+++ payloads/v10/channel.ts
@@ -8,6 +8,7 @@ import type { APIPartialEmoji } from './emoji';
 import type { APIGuildMember } from './guild';
 import type { APIInteractionDataResolved, APIMessageInteraction, APIMessageInteractionMetadata } from './interactions';
 import type { APIRole } from './permissions';
+import type { APIPoll } from './poll';
 import type { APISticker, APIStickerItem } from './sticker';
 import type { APIUser } from './user';
 
@@ -721,6 +722,17 @@ export interface APIMessage {
 	 * See https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object-resolved-data-structure
 	 */
 	resolved?: APIInteractionDataResolved;
+	/**
+	 * A poll!
+	 *
+	 * The `MESSAGE_CONTENT` privileged gateway intent is required for verified applications to receive a non-empty value from this field
+	 *
+	 * In the Discord Developers Portal, you need to enable the toggle of this intent of your application in **Bot > Privileged Gateway Intents**.
+	 * You also need to specify the intent bit value (`1 << 15`) if you are connecting to the gateway
+	 *
+	 * See https://support-dev.discord.com/hc/articles/4404772028055
+	 */
+	poll?: APIPoll;
 }
 
 /**